ci: disable features by GitHub API in cla workflow as a workaround for Github limitations
Due to the GitHub limitations, a github action including the Github API write call cannot be executed from the forked repo.
So, in that case, we cannot add labels or add comments on the issue, or do anything which requires write permissions.
In the future, we should make it possible to do the above things by introducing Github Apps.