akgraphics: use render pts instead of audio's current time for syncing
For AV-Sync, we currently sync a video frame twice.
The first time occurs in the mainloop in which we sync the pts of the frame to the audio's current time.
The second time occurs when rendering a video frame in which we selected the video frame based on the audio's current time.

This commit changes the behavior in the second time.
The video frame will be selected based on the pts of the frame to be rendered.