akcodec: use a safer way for convering PTS
There exists a high risk of numerical errors in the current method because the value of `AV_TIME_BASE` is relatively high.
For this change, instead of `AV_TIME_BASE`, we will use the original timebase whose value is expected to be lower than the value of it in most cases.