akcodec: fix a bug on resampling to planer format
In the current implementation, resampling to planer format does not work correctly, and its output has only one channel data.
This is because that output buffer is a plane pointer, not an array of pointer.

With this change, output buffer will be defined as an array of pointer.
And when resampling, the output buffer will be allocated properly by av_samples_alloc().