akcodec: add `is_active` field to `DecodeStream` to distinguish if it's actually being used
Currently, we have included all the streams in `InputSource`, and that causes erroneous behaviors you see in FFLayerSource::is_streams_end(), for example.

This commit introduces a new field `is_active` to `DecodeStream` to distinguish if the stream is actually being used, so that we can avoid issues above without breaking backward compatibility too much.