akeval(not-resolved): fix an issue that layers stay rendered even after set to display=false
This commit fixes the problem above, but it might not be a good solution.
Even with this change, the layer (display=False) stil remains in `m_targets` of akgraphics::RenderScene,
and are wasting memory and performance as it keeps its resources like textures or something till its atom be destroyed.

Since this design has existed since at least when support-enc be merged, this commit does not change the design for backward compatibility.