engine: fix or workaround the black screen issue occured when seeked to near the last frame
When seeked to near the last frame and then jump back seek executed, you will have a black screen on the monitor.
This is because in that case there is a discrepancy between the loop_cnt referenced by the audio callback and and the loop_cnt referenced by the decoder.
To fix that, this commit added an adjustment within DecodeLoop::decode_thread() that resolves the discrepancy between them.