akcodec: fix the initialization of `effective_pts`
Currently, we initialize `effective_pts` by the value of `AVFrame::pts`. This implentation is based on the assumption that the time base of the audio frame is always equal to the inverse of the sampling rate, which is the time base of `effective_pts`. However, this is not always the case (as in the case of MP3).

This change will avoid this pitfall by explicitly converting the time base of the `AVFrame::pts`.